Updated dependencies. Refactored landing, about, privacy and terms pages to be organized under a single directory,
