+++ fervent-web/src/components/about/LandingPage.jsx
@@ -2,13 +2,13 @@ import React, { useContext, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import Box from '@mui/material/Box';
 import Divider from '@mui/material/Divider';
-import AppAppBar from './components/AppAppBar';
-import Hero from './components/Hero';
-import Highlights from './components/Highlights';
-import Pricing from './components/Pricing';
-import Features from './components/Features';
-import FAQ from './components/FAQ';
-import Footer from './components/Footer';
+import AppAppBar from './components/landing-page/AppAppBar';
+import Hero from './components/landing-page/Hero';
+import Highlights from './components/landing-page/Highlights';
+import Pricing from './components/landing-page/Pricing';
+import Features from './components/landing-page/Features';
+import FAQ from './components/landing-page/FAQ';
+import Footer from './components/landing-page/Footer';
 import { AppContext } from '../context-provider/AppContext';
 import { appDescription, appName, metaDescription, metaKeywords } from '../../config/appConfig';
