+++ fervent-web/src/components/about/components/about-page/Introduction.jsx
@@ -4,7 +4,7 @@ import Box from '@mui/material/Box';
 import Container from '@mui/material/Container';
 import Stack from '@mui/material/Stack';
 import Typography from '@mui/material/Typography';
-import { orgName } from '../../../config/appConfig';
+import { orgName } from '../../../../config/appConfig';
 
 function Introduction() {
   return (
