+++ fervent-web/src/components/about/components/landing-page/Hero.jsx
@@ -6,8 +6,8 @@ import Container from '@mui/material/Container';
 import Link from '@mui/material/Link';
 import Stack from '@mui/material/Stack';
 import Typography from '@mui/material/Typography';
-import bgImage from '../assets/supported-platforms.svg';
-import { appName } from '../../../config/appConfig';
+import bgImage from '../../assets/supported-platforms.svg';
+import { appName } from '../../../../config/appConfig';
 
 const parts = appName.split(" ");
 const firstPart = `${parts[0]} ${parts[1]}`;
